+++ tests/testthat/test-plot.R
@@ -36,7 +36,7 @@ test_that("ggplot2 plots work as expected", suppressWarnings({
   skip_if_not_installed("vdiffr")
   skip_if_not_installed("ggplot2")
   skip_if_not_installed("errors", "0.3.6.1")
-  skip_if_not_installed("units", "0.8-0")
+  skip_if_not_installed("units", "0.8-5.5")
   library(ggplot2)
 
   iris.q <- iris