+++ NEWS.md
@@ -1,6 +1,7 @@
 # quantities devel
 
 - Performance improvements for `data.frame` methods.
+- Update `ggplot2` tests (r-quantities/units#367).
 
 # quantities 0.2.1